requirements.txt
python-dotenv
aiogram
groqcloud==0.4.0